Resolve crashpad permission error
When using headless, if the launching OS process doesn't have permission
in the directory that Chrome is found in the crashpad reporter will
attempt to `mkdir` and error out. This did not prevent Chrome from
launching, it would just continuously log out and be noisy.

This PR addresses this by setting the crash dump directory to a tmp one
that is writable by the launching process.

This flag I believe is only necessary on `darwin` and should be ignored
on other OSes